% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project2.R
\name{create_project2}
\alias{create_project2}
\title{Create Project 2}
\usage{
create_project2(new_project, ...)
}
\arguments{
\item{new_project}{new_project is a character vector with the file path and name of the project}

\item{...}{the ... argument takes optional character vector names for the additional folders to be added to the project structure}
}
\value{
does not return anything
}
\description{
create_project2() leverages usethis::create_project() with a few customizations. It has a standard project structure that's more complete that create_project(). In addition, if the user passes on optional folders, those will be added to the project structure. By default, the project will contain this folders: \code{data}, \code{images}, \code{plots}, and \code{scripts}
}
\examples{
\dontrun{
a new project called "my-project" with the "standard" project structure:
`create_project2("~/Desktop/my-project")`

a new project called "my-project" with an additional folder for reports
`create_project2("~/Desktop/my-project", "reports")`
}
}
