% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eltoolkit-package.R
\docType{package}
\name{eltoolkit-package}
\alias{eltoolkit}
\alias{eltoolkit-package}
\title{eltoolkit: Erwin Lares' Toolkit for R}
\description{
A collection of functions that presumably makes your life easier when working with R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erwinlares/eltoolkit}
  \item \url{https://erwinlares.github.io/eltoolkit/}
  \item Report bugs at \url{https://github.com/erwinlares/eltoolkit/issues}
}

}
\author{
\strong{Maintainer}: Erwin Lares \email{erwin.lares@wisc.edu} (\href{https://orcid.org/000000023284828X}{ORCID})

}
\keyword{internal}
